+++ apps/meteor/client/views/room/composer/messageBox/MessageBoxFormattingToolbar/MessageBoxFormattingToolbar.tsx
@@ -7,13 +7,13 @@ import type { ComposerAPI } from '../../../../../lib/chats/ChatAPI';
 import FormattingToolbarDropdown from './FormattingToolbarDropdown';
 
 type MessageBoxFormattingToolbarProps = {
-	disabled: boolean;
-	items: FormattingButton[];
 	composer: ComposerAPI;
 	variant?: 'small' | 'large';
+	items: FormattingButton[];
+	disabled: boolean;
 };
 
-const MessageBoxFormattingToolbar = ({ items, variant = 'large', composer, ...props }: MessageBoxFormattingToolbarProps) => {
+const MessageBoxFormattingToolbar = ({ items, variant = 'large', composer, disabled }: MessageBoxFormattingToolbarProps) => {
 	const t = useTranslation();
 
 	if (variant === 'small') {
@@ -24,12 +24,12 @@ const MessageBoxFormattingToolbar = ({ items, variant = 'large', composer, ...pr
 			<>
 				{'icon' in featuredFormatter && (
 					<MessageComposerAction
-						{...props}
 						onClick={() => composer.wrapSelection(featuredFormatter.pattern)}
 						icon={featuredFormatter.icon}
+						disabled={disabled}
 					/>
 				)}
-				<FormattingToolbarDropdown {...props} composer={composer} items={collapsedItems} />
+				<FormattingToolbarDropdown composer={composer} items={collapsedItems} disabled={disabled} />
 			</>
 		);
 	}
@@ -39,7 +39,7 @@ const MessageBoxFormattingToolbar = ({ items, variant = 'large', composer, ...pr
 			{items.map((formatter) =>
 				'icon' in formatter ? (
 					<MessageComposerAction
-						{...props}
+						disabled={disabled}
 						icon={formatter.icon}
 						key={formatter.label}
 						data-id={formatter.label}
@@ -54,8 +54,7 @@ const MessageBoxFormattingToolbar = ({ items, variant = 'large', composer, ...pr
 					/>
 				) : (
 					<span
-						{...props}
-						{...(props.disabled && { style: { pointerEvents: 'none' } })}
+						{...(disabled && { style: { pointerEvents: 'none' } })}
 						className='rc-message-box__toolbar-formatting-item'
 						title={formatter.label}
 						key={formatter.label}