+++ apps/meteor/server/models/raw/LoginServiceConfiguration.ts
@@ -1,11 +1,11 @@
-import type { ILoginServiceConfiguration, RocketChatRecordDeleted } from '@rocket.chat/core-typings';
+import type { LoginServiceConfiguration, RocketChatRecordDeleted } from '@rocket.chat/core-typings';
 import type { ILoginServiceConfigurationModel } from '@rocket.chat/model-typings';
 import type { Collection, Db } from 'mongodb';
 
 import { BaseRaw } from './BaseRaw';
 
-export class LoginServiceConfigurationRaw extends BaseRaw<ILoginServiceConfiguration> implements ILoginServiceConfigurationModel {
-	constructor(db: Db, trash?: Collection<RocketChatRecordDeleted<ILoginServiceConfiguration>>) {
+export class LoginServiceConfigurationRaw extends BaseRaw<LoginServiceConfiguration> implements ILoginServiceConfigurationModel {
+	constructor(db: Db, trash?: Collection<RocketChatRecordDeleted<LoginServiceConfiguration>>) {
 		super(db, 'meteor_accounts_loginServiceConfiguration', trash, {
 			preventSetUpdatedAt: true,
 			collectionNameResolver(name) {