+++ apps/meteor/tests/e2e/page-objects/home-channel.ts
@@ -40,4 +40,12 @@ export class HomeChannel {
 		await this.toastSuccess.locator('button >> i.rcx-icon--name-cross.rcx-icon').click();
 		await this.page.mouse.move(0, 0);
 	}
+
+	get composerToolbar(): Locator {
+		return this.page.locator('[role=toolbar][aria-label="Composer Primary Actions"]');
+	}
+
+	get composerToolbarActions(): Locator {
+		return this.page.locator('[role=toolbar][aria-label="Composer Primary Actions"] button');
+	}
 }