+++ apps/meteor/tests/end-to-end/api/18-oauthapps.js
@@ -63,6 +63,32 @@ describe('[OAuthApps]', function () {
 				})
 				.end(done);
 		});
+		it('should return a 403 Forbidden error when the user does not have the necessary permission by client id', (done) => {
+			updatePermission('manage-oauth-apps', []).then(() => {
+				request
+					.get(api('oauth-apps.get?clientId=zapier'))
+					.set(credentials)
+					.expect(403)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+						expect(res.body.error).to.be.equal('unauthorized');
+					})
+					.end(done);
+			});
+		});
+		it('should return a 403 Forbidden error when the user does not have the necessary permission by app id', (done) => {
+			updatePermission('manage-oauth-apps', []).then(() => {
+				request
+					.get(api('oauth-apps.get?appId=zapier'))
+					.set(credentials)
+					.expect(403)
+					.expect((res) => {
+						expect(res.body).to.have.property('success', false);
+						expect(res.body.error).to.be.equal('unauthorized');
+					})
+					.end(done);
+			});
+		});
 	});
 
 	describe('[/oauth-apps.create]', () => {