+++ packages/core-typings/src/IMessage/IMessage.ts
@@ -13,11 +13,11 @@ import type { IUser } from '../IUser';
 import type { FileProp } from './MessageAttachment/Files/FileProp';
 import type { MessageAttachment } from './MessageAttachment/MessageAttachment';
 
-type MessageUrl = {
+export type MessageUrl = {
 	url: string;
 	source?: string;
 	meta: Record<string, string>;
-	headers?: { contentLength: string } | { contentType: string } | { contentLength: string; contentType: string };
+	headers?: { contentLength?: string; contentType?: string };
 	ignoreParse?: boolean;
 	parsedUrl?: Pick<UrlWithStringQuery, 'host' | 'hash' | 'pathname' | 'protocol' | 'port' | 'query' | 'search' | 'hostname'>;
 };
@@ -357,16 +357,22 @@ export type IE2EEMessage = IMessage & {
 	e2e: 'pending' | 'done';
 };
 
-export type IOTRMessage = IMessage & {
-	t: 'otr' | 'otr-ack';
-};
+export interface IOTRMessage extends IMessage {
+	t: 'otr';
+	otrAck?: string;
+}
+
+export interface IOTRAckMessage extends IMessage {
+	t: 'otr-ack';
+}
 
 export type IVideoConfMessage = IMessage & {
 	t: 'videoconf';
 };
 
 export const isE2EEMessage = (message: IMessage): message is IE2EEMessage => message.t === 'e2e';
-export const isOTRMessage = (message: IMessage): message is IOTRMessage => message.t === 'otr' || message.t === 'otr-ack';
+export const isOTRMessage = (message: IMessage): message is IOTRMessage => message.t === 'otr';
+export const isOTRAckMessage = (message: IMessage): message is IOTRAckMessage => message.t === 'otr-ack';
 export const isVideoConfMessage = (message: IMessage): message is IVideoConfMessage => message.t === 'videoconf';
 
 export type IMessageWithPendingFileImport = IMessage & {