+++ ee/apps/ddp-streamer/CHANGELOG.md
@@ -1,5 +1,43 @@
 # @rocket.chat/ddp-streamer
 
+## 0.2.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+- @rocket.chat/rest-typings@6.5.3
+- @rocket.chat/core-services@0.3.3
+- @rocket.chat/model-typings@0.2.3
+- @rocket.chat/ui-contexts@3.0.3
+- @rocket.chat/models@0.0.27
+- @rocket.chat/instance-status@0.0.27
+
+## 0.2.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+- @rocket.chat/rest-typings@6.5.2
+- @rocket.chat/core-services@0.3.2
+- @rocket.chat/model-typings@0.2.2
+- @rocket.chat/ui-contexts@3.0.2
+- @rocket.chat/models@0.0.26
+- @rocket.chat/instance-status@0.0.26
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/rest-typings@6.5.1
+  - @rocket.chat/core-typings@6.5.1
+  - @rocket.chat/core-services@0.3.1
+  - @rocket.chat/ui-contexts@3.0.1
+  - @rocket.chat/model-typings@0.2.1
+  - @rocket.chat/models@0.0.25
+  - @rocket.chat/instance-status@0.0.25
+
 ## 0.2.0
 
 ### Minor Changes