+++ packages/model-typings/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @rocket.chat/model-typings
 
+## 0.2.3
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.3
+
+## 0.2.2
+
+### Patch Changes
+
+- @rocket.chat/core-typings@6.5.2
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [c2b224fd82]
+  - @rocket.chat/core-typings@6.5.1
+
 ## 0.2.0
 
 ### Minor Changes